refactor  goapp win
